Add EVENT_TYPE_INPUT_KEYPRESS_RESUMED to fix DeadKeyPreCombiner bug
When typing q [mark] w [mark] e [mark], the mark gets applied on the e but remains as separate combining characters on q and w. When the word finishes composing but gets composed again, the marks get reprocessed and they disappear. Avoid processing combining characters if they did not come from a keypress