Call imeManager.onStartInput() before other logic that can touch IMEs
LatinIMELegacy asked GeneralIME for autocaps flags when onStartInputView was called, but this call occurred before GeneralIME had onStartInput called. So it wasn't aware yet that input has restarted and caches need to be cleared, leading to incorrect autocaps flags. This wasn't usually a problem because a subsequent update was soon made in resetComposingWord -> resetEntireInputState, but under some configurations (e.g. disabled autocorrect and predictions) this didn't happen.